Use f32 instead of f64

+++ feynsum-rust/src/types.rs
@@ -2,11 +2,11 @@ use num::complex;
 
 pub type QubitIndex = usize;
 pub type GateIndex = usize;
-pub type Real = f64;
+pub type Real = f32;
 pub type Complex = complex::Complex<Real>;
 
 pub mod constants {
-    pub const RECP_SQRT_2: super::Real = std::f64::consts::FRAC_1_SQRT_2;
+    pub const RECP_SQRT_2: super::Real = std::f32::consts::FRAC_1_SQRT_2;
     pub const ZERO_THRESHOLD: super::Real = 0.00000001;
 }
